[Fix #160295024] Restructures the apps directory structure Removes app from API_V1 folder to the parent folder moves routes to views folder in app/a_p_i/v1 path moves models to a_p_1/v1 directory moves utility to app/a_p_i directory Reconfigures .travis.yml config file
