+++ .travis.yml
@@ -11,8 +11,8 @@ install:
 
 #run tests with
 script:
-  - cd API_V1/tests/v1
-  - nosetests  --with-coverage --cover-package=resources
+  - cd app/a_p_i/v1/tests
+  - nosetests  --with-coverage --cover-package=app/a_p_i/v1/views
 
 # command to run test coverage
 after_success:
